feat: cook book creation
